add sieve of atkin solution in scala

+++ problems/003/scala/sundaram.scala
@@ -1,8 +1,8 @@
 val target: Long = 600851475143L
 
 // http://en.wikipedia.org/wiki/Sieve_of_Sundaram
-def sundaramSieve(max: Long): Seq[Long] = {
-  val range = (2L to max)
+def sundaramSieve(max: Int): Seq[Int] = {
+  val range = (2 to max)
   range.diff(for {
     i <- range
     j <- range
@@ -10,7 +10,7 @@ def sundaramSieve(max: Long): Seq[Long] = {
   } yield candidate).map(_ * 2 + 1)
 }
 
-val primes = sundaramSieve(10000L)
+val primes = sundaramSieve(10000)
 val primeFactors = primes.filter(target % _ == 0)
 val largestPrime = primeFactors.reverse.head
